# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /finance/accounts:
    get:
      operationId: FinanceController_listAccounts
      parameters: []
      responses:
        '200':
          description: Returns all accounts for the authenticated user.
      security: &ref_7
        - bearer: []
      summary: List accounts
      tags: &ref_8
        - Finance
    post:
      operationId: FinanceController_createAccount
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountDto'
      responses:
        '201':
          description: Account created successfully.
      security: *ref_7
      summary: Create an account
      tags: *ref_8
  /finance/accounts/{id}:
    get:
      operationId: FinanceController_getAccount
      parameters:
        - name: id
          required: true
          in: path
          description: Account identifier
          schema:
            type: number
      responses:
        '200':
          description: Returns the requested account.
      security: *ref_7
      summary: Get account by id
      tags: *ref_8
    patch:
      operationId: FinanceController_updateAccount
      parameters:
        - name: id
          required: true
          in: path
          description: Account identifier
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountDto'
      responses:
        '200':
          description: Account updated successfully.
      security: *ref_7
      summary: Update an account
      tags: *ref_8
    delete:
      operationId: FinanceController_deleteAccount
      parameters:
        - name: id
          required: true
          in: path
          description: Account identifier
          schema:
            type: number
      responses:
        '200':
          description: Account deleted successfully.
      security: *ref_7
      summary: Delete an account
      tags: *ref_8
  /finance/categories:
    get:
      operationId: FinanceController_listCategories
      parameters: []
      responses:
        '200':
          description: Returns all categories for the authenticated user.
      security: *ref_7
      summary: List categories
      tags: *ref_8
    post:
      operationId: FinanceController_createCategory
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Category created successfully.
      security: *ref_7
      summary: Create a category
      tags: *ref_8
  /finance/categories/tree:
    get:
      operationId: FinanceController_categoryTree
      parameters: []
      responses:
        '200':
          description: Returns categories organised as a hierarchy.
      security: *ref_7
      summary: Get category tree
      tags: *ref_8
  /finance/categories/{id}:
    patch:
      operationId: FinanceController_updateCategory
      parameters:
        - name: id
          required: true
          in: path
          description: Category identifier
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Category updated successfully.
      security: *ref_7
      summary: Update a category
      tags: *ref_8
    delete:
      operationId: FinanceController_deleteCategory
      parameters:
        - name: id
          required: true
          in: path
          description: Category identifier
          schema:
            type: number
      responses:
        '200':
          description: Category deleted successfully.
      security: *ref_7
      summary: Delete a category
      tags: *ref_8
  /finance/transactions:
    get:
      operationId: FinanceController_listTransactions
      parameters:
        - name: type
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - INCOME
                - EXPENSE
                - TRANSFER
        - name: accountId
          required: false
          in: query
          schema:
            type: number
        - name: categoryId
          required: false
          in: query
          schema:
            type: number
        - name: dateFrom
          required: false
          in: query
          description: ISO date-time; accepts dateFrom or from query param
          schema:
            format: date-time
            type: string
        - name: dateTo
          required: false
          in: query
          description: ISO date-time; accepts dateTo or to query param
          schema:
            format: date-time
            type: string
        - name: minAmount
          required: false
          in: query
          schema:
            example: 10
            type: number
        - name: maxAmount
          required: false
          in: query
          schema:
            example: 500
            type: number
        - name: tags
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
        - name: q
          required: false
          in: query
          description: Full text search by notes, attachments, or related names
          schema:
            type: string
      responses:
        '200':
          description: Returns a paginated list of transactions.
      security: *ref_7
      summary: List transactions
      tags: *ref_8
    post:
      operationId: FinanceController_createTransaction
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionDto'
      responses:
        '201':
          description: Transaction created successfully.
      security: *ref_7
      summary: Create a transaction
      tags: *ref_8
  /finance/transactions/{id}:
    get:
      operationId: FinanceController_getTransaction
      parameters:
        - name: id
          required: true
          in: path
          description: Transaction identifier
          schema:
            type: number
      responses:
        '200':
          description: Returns the requested transaction.
      security: *ref_7
      summary: Get transaction by id
      tags: *ref_8
    patch:
      operationId: FinanceController_updateTransaction
      parameters:
        - name: id
          required: true
          in: path
          description: Transaction identifier
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionDto'
      responses:
        '200':
          description: Transaction updated successfully.
      security: *ref_7
      summary: Update a transaction
      tags: *ref_8
    delete:
      operationId: FinanceController_deleteTransaction
      parameters:
        - name: id
          required: true
          in: path
          description: Transaction identifier
          schema:
            type: number
      responses:
        '200':
          description: Transaction deleted successfully.
      security: *ref_7
      summary: Delete a transaction
      tags: *ref_8
  /finance/budgets:
    get:
      operationId: FinanceController_listBudgets
      parameters:
        - name: year
          required: false
          in: query
          schema:
            example: 2025
            type: number
        - name: month
          required: false
          in: query
          schema:
            example: 3
            type: number
      responses:
        '200':
          description: Returns budgets for the authenticated user.
      security: *ref_7
      summary: List budgets
      tags: *ref_8
    post:
      operationId: FinanceController_createBudget
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBudgetDto'
      responses:
        '201':
          description: Budget created successfully.
      security: *ref_7
      summary: Create a budget
      tags: *ref_8
  /finance/budgets/{id}:
    get:
      operationId: FinanceController_getBudget
      parameters:
        - name: id
          required: true
          in: path
          description: Budget identifier
          schema:
            type: number
      responses:
        '200':
          description: Returns the requested budget.
      security: *ref_7
      summary: Get budget by id
      tags: *ref_8
    patch:
      operationId: FinanceController_updateBudget
      parameters:
        - name: id
          required: true
          in: path
          description: Budget identifier
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBudgetDto'
      responses:
        '200':
          description: Budget updated successfully.
      security: *ref_7
      summary: Update a budget
      tags: *ref_8
    delete:
      operationId: FinanceController_deleteBudget
      parameters:
        - name: id
          required: true
          in: path
          description: Budget identifier
          schema:
            type: number
      responses:
        '200':
          description: Budget deleted successfully.
      security: *ref_7
      summary: Delete a budget
      tags: *ref_8
...
info:
  title: Finace API
  description: OpenAPI documentation for the Finace backend services.
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    jwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Provide the access token obtained from the auth endpoints.
  schemas:
...
    CreateAccountDto:
      type: object
      properties:
        name:
          type: string
          example: Main Checking
        type:
          type: string
          example: checking
          description: Account type label
        initialBalance:
          type: number
          example: 1000.5
      required:
        - name
        - type
    UpdateAccountDto:
      type: object
      properties:
        name:
          type: string
          example: Main Checking
        type:
          type: string
          example: checking
          description: Account type label
        initialBalance:
          type: number
          example: 1500.75
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: string
          example: Utilities
        parentId:
          type: object
          description: Parent category id
      required:
        - name
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: string
          example: Utilities
        parentId:
          type: object
          description: Parent category id
          nullable: true
    CreateTransactionDto:
      type: object
      properties:
        type:
          type: string
          enum: &ref_11
            - INCOME
            - EXPENSE
            - TRANSFER
        accountId:
          type: number
          example: 1
          description: Account identifier
        targetAccountId:
          type: number
          example: 2
          description: Target account for transfers
        categoryId:
          type: number
          example: 1
          description: Category identifier
        amount:
          type: number
          example: 125.75
        occurredAt:
          type: string
          format: date-time
        notes:
          type: string
          maxLength: 500
        tags:
          example: &ref_12
            - groceries
            - family
          type: array
          items:
            type: string
        attachmentUrl:
          type: string
          example: https://example.com/receipt.pdf
        installmentsTotal:
          type: number
          example: 12
        installmentNumber:
          type: number
          example: 1
        recurrenceRRule:
          type: string
          description: RFC5545 RRULE for recurring transactions
      required:
        - type
        - accountId
        - amount
        - occurredAt
    UpdateTransactionDto:
      type: object
      properties:
        type:
          type: string
          enum: *ref_11
        accountId:
          type: number
          example: 1
          description: Account identifier
        targetAccountId:
          type: number
          example: 2
          description: Target account for transfers
        categoryId:
          type: number
          example: 1
          description: Category identifier
        amount:
          type: number
          example: 300.25
        occurredAt:
          type: string
          format: date-time
        notes:
          type: string
          maxLength: 500
        tags:
          example: *ref_12
          type: array
          items:
            type: string
        attachmentUrl:
          type: string
          example: https://example.com/receipt.pdf
        installmentsTotal:
          type: number
          example: 12
        installmentNumber:
          type: number
          example: 1
        recurrenceRRule:
          type: string
          description: RFC5545 RRULE for recurring transactions
    CreateBudgetDto:
      type: object
      properties:
        categoryId:
          type: number
          example: 1
        year:
          type: number
          example: 2025
        month:
          type: number
          example: 3
        plannedAmount:
          type: number
          example: 1500
      required:
        - categoryId
        - year
        - month
        - plannedAmount
    UpdateBudgetDto:
      type: object
      properties:
        categoryId:
          type: number
          example: 1
        year:
          type: number
          example: 2025
        month:
          type: number
          example: 5
        plannedAmount:
          type: number
          example: 1600